Gráficos Melhorados

+++ scripts/puxa.js
@@ -1,113 +1,19 @@
 let buttonControlGraphic, stateBt = false;
-let vet = [];
-let chart;
-let vetFim = []; //data
-
+let typeTime = null;
+const typesLast = ["ult5min", "ult30min", "ult1h", "ult2h", "ult5h", "ult10h", "ult24h"]
 let btConsulta;
 
+let stateSlctAll = false;
+
+let btSelectAll = document.querySelector('#slct-all');
 
 
-
-let options = {
-    chart: {
-        type: 'line',
-        events: {
-            load: async function () {
-
-                // set up the updating of the chart each second
-                var series = this.series[0];
-                setInterval(function () {
-                    var x = (new Date()).getTime() // current time
-                    //let dados = puxaData();
-                    var y = Math.round(Math.random() * 100);
-                    series.addPoint([x, y], true, true);
-                }, 5000);
-            }
-        }
-    },
-
-    time: {
-        useUTC: false
-    },
-
-    yAxis: {
-        title: {
-            text: "Temperatura (C)"
-        },
-        pointStart: 0,
-        plotlines: [{ value: 0, width: 1, "color": "#808080" }]
-    },
-    tooltip: {
-        valueSuffix: "C°"
-    },
-    legend: {
-        layout: "vertical",
-        align: "right",
-        verticalAlign: "middle",
-        borderWidth: 0
-    },
-
-    title: {
-        text: 'Live random data'
-    },
-
-    exporting: {
-        enabled: false
-    },
-
-    series: [{
-        name: 'T1',
-        data: vetFim
-    }],
-    plotOptions: {
-        bar: {
-            colorByPoint: true
-        },
-        series: {
-            zones: [{
-                color: '#4CAF50',
-                value: 0
-            }, {
-                color: '#8BC34A',
-                value: 10
-            }, {
-                color: '#CDDC39',
-                value: 20
-            }, {
-                color: '#CDDC39',
-                value: 30
-            }, {
-                color: '#FFEB3B',
-                value: 40
-            }, {
-                color: '#FFEB3B',
-                value: 50
-            }, {
-                color: '#FFC107',
-                value: 60
-            }, {
-                color: '#FF9800',
-                value: 70
-            }, {
-                color: '#FF5722',
-                value: 80
-            }, {
-                color: '#F44336',
-                value: 90
-            }, {
-                color: '#F44336',
-                value: Number.MAX_VALUE
-            }]
-        }
-    }
-};
-
-var urlencoded = new URLSearchParams();
-urlencoded.append("option", "tudo");
+let resolution = null;
 
 
-let onChangeGraph;
-
+const rotinaLimpezaObjetoChart = () => {
+    options.series = [];
+}
 
 
 const deleteGraph = () => {
@@ -117,6 +23,7 @@ const deleteGraph = () => {
     }
     $('#chart').remove(); 
     $("div.panel-body").append('<div id="chart"></div>');
+    rotinaLimpezaObjetoChart(); // clear obj options of chart
 }
 
 $(document).ready(function () {
@@ -148,56 +55,15 @@ $(document).ready(function () {
     //
 
     btConsulta = document.querySelectorAll('#button-realiza-consulta')[0];
-    console.log(btConsulta);
+
     btConsulta.addEventListener('click', doRequestQueryAPI);
 
     document.querySelectorAll('#button-reset-inputs')[0].addEventListener('click', resetAllInputs);
 
-    puxaData();
 });
 
-const puxaData = async () => {
-    var myHeaders = new Headers();
-    myHeaders.append("Content-Type", "application/x-www-form-urlencoded");
-
-
-
-    var requestOptions = {
-        method: 'POST',
-        headers: myHeaders,
-        body: urlencoded,
-        redirect: 'follow'
-    };
-
-    let fim = await fetch("http://concentradorsolar.000webhostapp.com/api", requestOptions)
-        .then(response => response.json())
-        .then((result) => {
-            console.log(result);
-            let resultInvert = result["content"].reverse();
-            resultInvert.forEach((element) => {
-                vet.push(parseInt(element[0].temperaturas[0]) + getRandomIntInclusive(1, 499));
-            });
-            var time = (new Date()).getTime();
-            vet.forEach((element) => {
-                vetFim.push([
-                    time + Math.random() * 100 ,
-                    Math.round(element + Math.random() * 100)
-                ]);
-            });
-            console.log(vetFim)
-            if (!chart) {
-                chart = new Highcharts.Chart('chart', options);
-            }
-        })
-        .catch(error => console.log('error', error));
-        
-    
-
-    return fim;
-}
-
 const onChangeDataSensor = () => {
-    const stateValue = ["temperatura", "vazao", "vento", "irradiancia"];
+    const stateValue = ["temperature", "flow", "velocityWind", "irradiance"];
     let equalElement = false;
     $('#slct').change(function () {
         checkAllInputsData();
@@ -207,32 +73,45 @@ const onChangeDataSensor = () => {
                 equalElement = true;
                 $('.quant-sensores-div').css('display', 'flex');
                 
-                if ($(this).val() == "temperatura") {
+                if ($(this).val() == "temperature") {
                     $('#temp-sensor').css('display', 'flex');
                     $('#vazao-sensor').css('display', 'none');
                     $('#vento-sensor').css('display', 'none');
                     $('#irradiancia-sensor').css('display', 'none');
+
+                    $('.select-all').css('display', 'flex');
+                    btSelectAll.addEventListener('click', selectAllTempSensors);
                 }
-                else if ($(this).val() == "vazao") {
+                else if ($(this).val() == "flow") {
                     $('#vazao-sensor').css('display', 'flex');
                     $('#temp-sensor').css('display', 'none');
                     $('#vento-sensor').css('display', 'none');
                     $('#irradiancia-sensor').css('display', 'none');
                     document.querySelectorAll('input#vazao')[0].checked = true;
+
+                    btSelectAll.removeEventListener('click', selectAllTempSensors);
+                    $('.select-all').css('display', 'none');
+                    
                 }
-                else if ($(this).val() == "vento") {
+                else if ($(this).val() == "velocityWind") {
                     $('#vento-sensor').css('display', 'flex');
                     $('#temp-sensor').css('display', 'none');
                     $('#vazao-sensor').css('display', 'none');
                     $('#irradiancia-sensor').css('display', 'none');
                     document.querySelectorAll('input#vento')[0].checked = true;
+
+                    btSelectAll.removeEventListener('click', selectAllTempSensors);
+                    $('.select-all').css('display', 'none');
                 }
-                else if ($(this).val() == "irradiancia"){
+                else if ($(this).val() == "irradiance"){
                     $('#irradiancia-sensor').css('display', 'flex');
                     $('#temp-sensor').css('display', 'none');
                     $('#vazao-sensor').css('display', 'none');
                     $('#vento-sensor').css('display', 'none');
                     document.querySelectorAll('input#irradiancia')[0].checked = true;
+
+                    btSelectAll.removeEventListener('click', selectAllTempSensors);
+                    $('.select-all').css('display', 'none');
                 }
                 else {
                     $('.quant-sensores-div').css('display', 'none');
@@ -246,6 +125,20 @@ const onChangeDataSensor = () => {
     });
 }
 
+function selectAllTempSensors() {
+    const nodes = ["t1", "t2", "t3", "t4", "t5", "t6", "t7", "t8", "t9", "t10"];
+    stateSlctAll = !stateSlctAll;
+
+    let typeSet = true;
+
+    if (!stateSlctAll) { // falso
+        typeSet = false;
+    }
+    nodes.forEach((element) => {
+        $(`#${element}`).prop('checked', typeSet);
+    })
+}
+
 function resetCheckButtons() {
     let nodes = document.querySelectorAll('input.inp-cbx');
 
@@ -257,8 +150,14 @@ function resetCheckButtons() {
 const onChangePeriodo = () => {
     const stateValue = ["ult5min", "ult30min", "ult1h", "ult2h", "ult5h", "ult10h", "ult24h", "custom"];
     let equalElement = false;
+
+    onChangeTimeCustom();
+
     $('#slct1').change(function () {
         checkAllInputsData();
+
+        typeTime = $('#slct1').val(); // guardar configuração de tipo de tempo
+
         stateValue.forEach((element) => {
             if (element == $(this).val()) {
                 equalElement = true;
@@ -291,6 +190,18 @@ const onInputDateSucced = () => {
     return diffDays;
 }
 
+const onChangeTimeCustom = () => {
+    let inputCustomType;
+    $('#customTime').change(() => {
+        inputCustomType = $('#customTime').is(":checked");
+        if (inputCustomType) {
+            $('.config-time-custom').css('display', 'flex');
+        }
+        else {
+            $('.config-time-custom').css('display', 'none');
+        }
+    })
+}
 
 const onChangeInputDate = () => {
     $('#initialDate, #endDate').change(function () {
@@ -299,16 +210,19 @@ const onChangeInputDate = () => {
             $('div.reso').css('display', 'flex');
             $('div.engloba-info').css('display', 'none');
             if(onInputDateSucced() == 1) {
+                resolution = $('#slct-date1').val();
                 $('.select-data-1').css('display', 'flex');
                 $('.select-data-2').css('display', 'none');
                 $('.select-data-3').css('display', 'none');
             }
             else if (onInputDateSucced() == 2) {
+                resolution = $('#slct-date2').val();
                 $('.select-data-2').css('display', 'flex');
                 $('.select-data-1').css('display', 'none');
                 $('.select-data-3').css('display', 'none');
             }
             else if (onInputDateSucced() == 3) {
+                resolution = $('#slct-date3').val();
                 $('.select-data-3').css('display', 'flex');
                 $('.select-data-2').css('display', 'none');
                 $('.select-data-1').css('display', 'none');
@@ -340,9 +254,7 @@ const onChangeInputTypeGraphic = () => {
 }
 
 const checkAllInputsData = () => {
-    if ($('#slct').val() !== null && $('#slct1').val() !== null && $('#slct2').val() !== null) {
-        console.log('aaa');
-        
+    if ($('#slct').val() !== null && $('#slct1').val() !== null && $('#slct2').val() !== null) {    
         $('div#button-realiza-consulta').removeClass('disabled');
         return true;
     }
@@ -354,8 +266,16 @@ const checkAllInputsData = () => {
 
 const doRequestQueryAPI = () => {
     if(checkAllInputsData()) {
-        deleteGraph();
-        puxaData();
+        if(chart) { // verifica se há necessidade de apagar caso exista um chart
+            deleteGraph();
+        }
+        
+        if (typeTime == "custom") {
+            puxaDataCustom();
+        }
+        else {
+            puxaLastData();
+        }
     }
 }
